% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{bayes_factor}
\alias{bayes_factor}
\title{Bayes factor}
\usage{
bayes_factor(SIR, prior_probs = NULL)
}
\arguments{
\item{SIR}{SIR Fit model or list of SIR fit models}

\item{prior_probs}{prior probabilities of each model}
}
\value{
vector of bayes factors
}
\description{
Bayes factor
}
