% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humpback_SIR_20171206.R
\name{CALC.ANALYTIC.Q}
\alias{CALC.ANALYTIC.Q}
\title{Compute analytic estimates of q, the scaling parameter between indices and
absolute population size}
\usage{
CALC.ANALYTIC.Q(rel.abundance, Pred_N, start_yr, add_CV = 0, num.IA)
}
\arguments{
\item{rel.abundance}{Relative abundance index}

\item{Pred_N}{Predicted population}

\item{start_yr}{Initial year}

\item{add_CV}{Coefficient of variation}

\item{num.IA}{Index of abundance}
}
\value{
A numeric estimator for $q$.
}
\description{
Compute analytic estimates of q, the scaling parameter between indices and
absolute population size
}
