% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SIR.R
\name{CALC.LNLIKE}
\alias{CALC.LNLIKE}
\title{Function to calculate the log-likelihood using a lognormal distribution}
\usage{
CALC.LNLIKE(Obs.N, Pred_N, CV, log = FALSE)
}
\arguments{
\item{Obs.N}{Time series of observed abundance}

\item{Pred_N}{Time series of estimated abundance}

\item{CV}{coefficient of variation}

\item{log}{whether to export as log-likelihood}
}
\value{
returns a scalar of the likelihood
}
\description{
Function to calculate the log-likelihood using a lognormal distribution
}
\examples{
Obs.N <- 2000
Pred_N <- 2340
CV <- 4
CALC.LNLIKE(Obs.N, Pred_N, CV)
}
