% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humpback_SIR_20171206.R
\name{COMPUTING.ROI}
\alias{COMPUTING.ROI}
\title{Computes the predicted rate of increase for a set of specified years for
comparison with trends estimated separately with any of the indices of
abundance or count data}
\usage{
COMPUTING.ROI(data = data, Pred_N = Pred_N, start_yr = NULL)
}
\arguments{
\item{data}{Count data or relative abundance index to use}

\item{Pred_N}{Number of individuals predicted}

\item{start_yr}{Initial year}
}
\value{
Vector of rates of increase, one per index
}
\description{
Computes the predicted rate of increase for a set of specified years for
comparison with trends estimated separately with any of the indices of
abundance or count data
}
