% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlnorm_zerb}
\alias{dlnorm_zerb}
\title{Adjusted lognormal likelihood from Zerbini et al. 2011 (eq. 5)}
\usage{
dlnorm_zerb(x, meanlog, sdlog, return_log = TRUE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{meanlog}{mean of the distribution on the log scale with a default value of 0.}

\item{sdlog}{standard deviation of the distribution on the log scale with a default value of 1.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A vector of densities.
}
\description{
\code{dlnorm_zerb} returns the density for the adjusted log normal distribution whose logarith has mean equal to \code{meanlog} and standard deviation equal to \code{sdlog}:
$$f\left(x\right) = \frac{1}{\sigma * x} exp \left( \frac{ \left( ln\left( x \right) - \mu \right) ^ 2} {2 \sigma^2} \right)$$
}
