% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getz}
\alias{getz}
\title{Get z from Nk}
\usage{
getz(Nk = 0.6, b = -1, maxiter = 100)
}
\arguments{
\item{Nk}{Depletion which maximizes yield.}

\item{b}{branch of lambert W function}

\item{maxiter}{Number of iterations to solve lambert W}
}
\description{
Get z from Nk
}
\examples{

getz(Nk = 0.6)
}
