% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humpback_SIR_20171206.R
\name{LNLIKE.BEACHED}
\alias{LNLIKE.BEACHED}
\title{Calculate the log-likelihood of the beach whale data}
\usage{
LNLIKE.BEACHED(beached_data, Pred_N, start_yr, q_anthro, d_anthro,
  p_anthro, log = TRUE)
}
\arguments{
\item{beached_data}{Data on whale strandingsas a data.frame
containing year, estimate of absolute abundance, and standard deviation}

\item{Pred_N}{Predicted absolute abundance in numbers from
\code{\link{GENERALIZED_LOGISTIC}}.}

\item{start_yr}{The first year of the projection (assumed to be the first
year in the catch series).}

\item{q_anthro}{is the proportion of the population that will be killed each year from anthropogenic mortality}

\item{d_anthro}{is the detection probability of carcasses on the beach (including the probability of washing up on shore)}

\item{p_anthro}{is the proportion of the range of the species covered by monitoring}

\item{log}{Return the log of the likelihood (TRUE/FALSE)}

\item{add_CV}{Additional CV to add to variance of lognormal distribution
sampled from \code{priors$add_CV}.}
}
\value{
the negative log-likelihood
}
\description{
Calculate the log-likelihood of the beach whale data
}
