% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humpback_SIR_20171206.R
\name{LNLIKE.GR}
\alias{LNLIKE.GR}
\title{Calculate the log-likelihood of the growth rate}
\usage{
LNLIKE.GR(Obs.GR, Pred.GR, GR.SD.Obs, log = T)
}
\arguments{
\item{Obs.GR}{Observed growth rate}

\item{Pred.GR}{Predicted growth rate}

\item{GR.SD.Obs}{Standard error of the observed growth rate}
}
\value{
A \code{list} containing \code{loglike.GR1} and \code{loglike.GR2}
}
\description{
Calculates the log-likelihood of the estimated growth rate given the observed
growth rate and the standard deviation of the observed growth rate.
}
\examples{
LNLIKE.GR(0.1, 0.1, 0.1)
}
