% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SIR.R
\name{LNLIKE.IAs}
\alias{LNLIKE.IAs}
\title{Compute the log-likelihood of indices of abundance}
\usage{
LNLIKE.IAs(rel.abundance, Pred_N, start_yr, q.values, add.CV, log = TRUE)
}
\arguments{
\item{rel.abundance}{Relative abundance}

\item{Pred_N}{Predicted population size}

\item{start_yr}{Initial year}

\item{q.values}{Scaling parameter}

\item{add.CV}{Coefficient of variation}

\item{log}{Boolean, return log likelihood (default TRUE) or
likelihood.}
}
\value{
List of likelihood based on Zerbini et al. (2011) eq. 5 or using `dnorm`
}
\description{
Compute the log-likelihood of indices of abundance
}
