% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SIR.R
\name{LNLIKE.Ns}
\alias{LNLIKE.Ns}
\title{LOG LIKELIHOOD OF ABSOLUTE ABUNDANCE}
\usage{
LNLIKE.Ns(Obs.N, Pred_N, start_yr, add_cv, log = TRUE)
}
\arguments{
\item{Obs.N}{Observed absoluted abundance in numbers as a data.frame
containing year, estimate of absolute abundance, and standard deviation}

\item{Pred_N}{Predicted absolute abundance in numbers from
\code{\link{GENERALIZED_LOGISTIC}}.}

\item{start_yr}{The first year of the projection (assumed to be the first
year in the catch series).}

\item{log}{Return the log of the likelihood (TRUE/FALSE)}

\item{add_CV}{Additional CV to add to variance of lognormal distribution
sampled from \code{priors$add_CV}.}
}
\value{
A list of two numeric scalars of estimates of log-likelihood.
}
\description{
This function computes two estimates of the log-likelihood of the estimated
absolute abundance using the equation from Zerbini et al. 2011 (eq. 4) and a
lognormal distribution from \code{\link{CALC.LNLIKE}}.
}
\examples{
Obs.N  <-  data.frame(Year = 2005, Sigma = 5, Obs.N = 1000)
Pred_N  <-  1234
start_yr  <-  2005
LNLIKE.Ns(Obs.N, Pred_N, start_yr, add_cv = 0, log=TRUE)
}
