% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SIR.R
\name{LOGISTIC.BISECTION.K}
\alias{LOGISTIC.BISECTION.K}
\title{LOGISTIC BISECTION}
\usage{
LOGISTIC.BISECTION.K(K.low, K.high, r_max, z, num_Yrs, start_yr,
  target.Pop, catches, MVP, tol = 0.001)
}
\arguments{
\item{K.low}{Lower bound for $K$ when preforming the bisection method of Punt
and Butterworth (1995). Default is 1.}

\item{K.high}{Upper bound for $K$ when preforming the bisection method of
Punt and Butterworth (1995). Default is 500,000.}

\item{r_max}{The maximum net recruitment rate ($r_{max}$).}

\item{z}{The parameter that determines the population size where productivity
is maximum (assumed to be 2.39 by the IWC SC).}

\item{num_Yrs}{The number of projection years. Set as the last year in the
catch or abundance series, whichever is most recent, minus the
\code{start_yr}.}

\item{start_yr}{The first year of the projection (assumed to be the first
year in the catch series).}

\item{target.Pop}{A sample of the prior on population abundance $N$, in
numbers, set as \code{sample.N.obs} sampled from \code{priors$N.obs}}

\item{catches}{The time series of catch in numbers or biomass. Currently does
not handle NAs and zeros will have to input a priori for years in which
there were no catches.}

\item{MVP}{The minimum viable population size in numbers or biomass. Computed
as 3 * \code{\link{num.haplotypes}} to compute minimum viable population
(from Jackson et al., 2006 and IWC, 2007).}

\item{tol}{The desired accuracy (convergence tolerance) of
\code{\link{stats::uniroot}}.}
}
\value{
A numeric scalar of an estimate of  carrying capacity $K$.
}
\description{
Method of Butterworth and Punt (1995) where the prior distribution of the
current absolute abundance $N_{2005}$ and maximum net recruitment rate
\code{r_max} are sampled and then used to determine the unique value of the
population abundance $N$ in \code{start_yr} (assumed to correspond to
carrying capacity $K$). Requires \code{\link{TARGET.K}} and subsequent
dependencies.
}
\examples{
LOGISTIC.BISECTION.K(K.low = 1, K.high = 100000, r_max = r_max, z = z,
                     num_Yrs = bisection.Yrs, start_yr = start_yr,
                     target.Pop = target.Pop, catches = catches, MVP = MVP,
                     tol = 0.001)
}
