% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{make_multiplier_list}
\alias{make_multiplier_list}
\title{Make a list of priors for the catch multiplier to be passed to the SIR function.}
\usage{
make_multiplier_list(c_mult_1 = make_prior(1), ...)
}
\arguments{
\item{c_mult_1}{Catch multiplier for the first catch period; defaults to 1 and used for the entire dataset}

\item{...}{Additional catch multipliers as defined using \code{make_prior}. Must be supplied in order of use.}
}
\value{
A named list containing each of the specified priors in a form that
  can be used by the SIR function.
}
\description{
Make a list of priors for the catch multiplier to be passed to the SIR function.
}
