% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_specification.R
\name{make_prior}
\alias{make_prior}
\title{Generate a prior for use in the Humpback SIR}
\usage{
make_prior(rfn = NA, par1 = NULL, par2 = NULL, use = TRUE,
  label = NULL)
}
\arguments{
\item{rfn}{Function to be used to generate samples from the prior
distribution, such as \code{rnorm} or \code{runif}. Can also be set to a
single number in order to hold that parameter constant.}

\item{par1}{First parameter of the prior distribution. For \code{rnorm} this
is the mean, for \code{runif} this is the lower bound.}

\item{par2}{Second parameter of the prior distribution. For \code{rnorm}
this is the standard deviation, for \code{runif} this is the upper bound.}

\item{use}{Boolean, indicates whether this prior is to be used. Defaults to
TRUE.}

\item{label}{Text label of the distribution. If not set, will try to label
based on the function passed in \code{rfn} Recognizes \code{rnorm}
\code{rlnorm} \code{runif}, and \code{rlunif}. Otherwise will label
"User defined".}
}
\value{
A \code{list} containing the function \code{rfn} for generating a
  sample from the prior distribution, a vector \code{pars} containing the
  parameters of the distribution, and a boolean \code{use} flag.
}
\description{
Convenient way to generate a prior that can be used to generate samples for
the SIR.
}
\examples{
make_prior(rnorm, 0, 1, TRUE)
make_prior(runif, 0, 1, TRUE)
make_prior(rlunif, 0.01, 0.2, "Log-uniform(0.01, 0.2)")
}
