% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sir.R
\name{plot_density}
\alias{plot_density}
\title{OUTPUT FUNCTION}
\usage{
plot_density(SIR, file_name = NULL, lower = NULL, upper = NULL,
  priors = NULL, inc_reference = TRUE, probs = c(0.025, 0.975))
}
\arguments{
\item{SIR}{A fit SIR model or list of SIR models. Plots in the order provided.}

\item{file_name}{name of a file to identified the files exported by the
function. If NULL, does not save.}

\item{lower}{Vector of lower bounds for x-axis}

\item{upper}{Vector of upper bounds for x-axis}

\item{priors}{Default = NULL, SIR realized priors or list of SIR realized priors. Plots in the order provided.}

\item{probs}{Lower and upper quantiles to use for plot limits if lower and upper are not specified.}

\item{reference}{Default = NULL, wether reference case is included in SIR}
}
\value{
Returns and saves a figure with the posterior densities of parameters.
}
\description{
Function that provides a plot of the estimated posterior densities of parameters from  SIR  model.
}
