% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sir.R
\name{plot_ioa}
\alias{plot_ioa}
\title{OUTPUT FUNCTION}
\usage{
plot_ioa(SIR, file_name = "NULL", ioa_names = NULL,
  posterior_pred = FALSE)
}
\arguments{
\item{SIR}{A fit SIR model}

\item{file_name}{name of a file to identified the files exported by the}

\item{ioa_names}{names of indices of abundance used.}

\item{posterior_pred}{Logical. If true, includes a posterior predictive distribution of the estimated IOA
function.}
}
\value{
Returns and saves a figure with the IOA trajectories.
}
\description{
Function that provides a plot of the estimated indices of abundance a SIR  model including: median, 95%
credible interval, 90% credible interval, catch, and observed indices of abundance abundance.
}
