% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sir.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{OUTPUT FUNCTION}
\usage{
plot_trajectory(SIR, Reference = NULL, file_name = "NULL")
}
\arguments{
\item{SIR}{A fit SIR model}

\item{Reference}{A fit SIR model for the reference case}

\item{file_name}{name of a file to identified the files exported by the
function.}
}
\value{
Returns and saves a figure with the population trajectory.
}
\description{
Function that provides a plot of the estimated population trajectory from a SIR outputs model including: median, 95%
credible interval, 90% credible interval, catch, and observed absolute abundance.
}
