% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SIR.R
\name{PRED.GROWTH.RATE}
\alias{PRED.GROWTH.RATE}
\title{PREDICTED GROWTH RATE}
\usage{
PRED.GROWTH.RATE(growth.rate.Yrs, Pred_N, start_yr = start_yr)
}
\arguments{
\item{growth.rate.Yrs}{The years to be used for growth rate computation. 1995 - 1998 are used in Zerbini et al. 2011.}

\item{Pred_N}{Time series of predicted abundance, in numbers, from \code{\link{GENERALIZED_LOGISTIC}}.}

\item{start_yr}{The first year of the projection (assumed to be the first year in the catch series).}
}
\value{
A numeric scalar representing predicted growth rate.
}
\description{
\code{PRED.GROWTH.RATE} computes the predicted growth rate if such
information is available from an independent estimate rather than being
estimated from data. Growth rate is calculated as: $$r_{t_0 - t_{fin}}^{pred}
= \frac{ \sum_{t = t_0} ^{t_{fin - 1}} ln \left( \frac{N_{t+1}^{pred}} {
N_t^{pred}} \right) } { t_{fin} - t_0 } = \frac{ ln \left( N_{fin}^{pred}
\right) - ln \left( N_{0}^{pred} \right)} { t_{fin} - t_0 }$$ where
$N^{pred}$ is the model predicted population size, in numbers, at time $t$ or
$t+1$ in years, $t_0$ is the start year of the equation (1995 in Zerbini et
al. 2011), and $t_{fin}$ is the last year of the equation (1998 in Zerbini et
al. 2011).
}
\examples{
growth.rate.Yrs  <-  c(1995:1998)
Pred_N <- c(1000, 1500, 1500, 2000)
start_yr  <-  1995
PRED.GROWTH.RATE(growth.rate.Yrs, Pred_N, start_yr=start_yr)
}
