% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humpback_SIR_20171206.R
\name{PREDICT.IAs}
\alias{PREDICT.IAs}
\title{Predict indices of abundance}
\usage{
PREDICT.IAs(rel.abundance, Pred_N, start_yr, q.values, add.CV,
  log = TRUE)
}
\arguments{
\item{SIR}{SIR object}
}
\value{
List of predicted indices based on Zerbini et al. (2011) eq. 5 or using `dnorm`
}
\description{
Predict indices of abundance
}
