% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_specification.R
\name{rbest}
\alias{rbest}
\title{Probability density for correction factor as determined by Best et al 2010}
\usage{
rbest(n = 1, crit_95 = 0.139, upper = 0.3)
}
\arguments{
\item{n}{number of samples}

\item{crit_95}{Struck and loss rate such that there is only a 5% probability of struck and loss rate exceeding this value}

\item{upper}{Upper limit for struck and loss rate}
}
\description{
Probability density for correction factor as determined by Best et al 2010
}
