% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_specification.R
\name{rlunif}
\alias{rlunif}
\title{The log-uniform distribution}
\usage{
rlunif(n, min = 1, max = 2)
}
\arguments{
\item{n}{Number of observations.}

\item{min, max}{Limits of the distribution (must be strictly positive). These
will be the limits of the samples returned.}
}
\value{
A vector of length \code{n} containing random draws from the
  log-uniform distribution.

Draws independent samples from Uniform(log(min), log(max)) and
exponentiates.
}
\description{
Return a random sample from the log-uniform distribution.
}
\examples{
rlunif(1, 0.01, 0.2)
}
