% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humpback_SIR_20171206.R
\name{SAMPLE.PRIOR}
\alias{SAMPLE.PRIOR}
\title{Function to sample from priors}
\usage{
SAMPLE.PRIOR(name = NA, Val.1 = NA, Val.2 = NA)
}
\arguments{
\item{name}{Name of a distribution, as a string}

\item{Val.1}{Lower bound of "(log-)uniform", mean of "(log-)normal"}

\item{Val.2}{Upper bound of "(log-)uniform", SD of "(log-)normal"}
}
\value{
Single random sample from the specified prior distribution.
}
\description{
Function to sample from priors
}
\examples{
SAMPLE.PRIOR("uniform", 0, 1)
SAMPLE.PRIOR("log-uniform", 1e-5, 10)
SAMPLE.PRIOR("normal", 0, 1)
SAMPLE.PRIOR("log-normal", 0, 1)
}
