% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humpback_SIR_20171206.R
\name{summary_sir}
\alias{summary_sir}
\title{OUTPUT FUNCTION}
\usage{
summary_sir(x, object = "USERDEFINED", file_name = "NULL")
}
\arguments{
\item{x}{A data.frame of mcmc samples.}

\item{object}{Name of the model object as specified by the user.}

\item{file_name}{name of a file to identified the files exported by the
function.}
}
\value{
Returns a data.frame with summary of SIR outputs
}
\description{
Function that provides a summary of SIR outputs including: mean, median, 95%
credible interval, 90% predicitive interval, max, and sample size.
}
\examples{
x  <-  rnorm(1000, 5, 7)
y  <-  rnorm(1000, 6, 9)
df <- data.frame(x = x, y = y)
summary_sir( df , object = "example_summary")
}
