% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{weight_model}
\alias{weight_model}
\title{Weighted SIR model}
\usage{
weight_model(SIR, bayes_factor)
}
\arguments{
\item{SIR}{List of SIR fit models}

\item{bayes_factor}{vector of bayes_factors from \code{\link{bayes_factor}}}
}
\value{
an object of class "SIR"
}
\description{
Function to create a weighted model using bayes factors
}
