% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{zerbini_table}
\alias{zerbini_table}
\title{Function to write tables of logistic model parameter and derived quantities for HumpbackSIR similar to Table 5 from Zerbini et al (2011).}
\usage{
zerbini_table(SIR, file_name = NULL)
}
\arguments{
\item{SIR}{Resample summary from hympbackSIR}

\item{file_name}{Desired filename to where csv file will be saved. If NULL, will not save.}
}
\description{
Function to write tables of logistic model parameter and derived quantities for HumpbackSIR similar to Table 5 from Zerbini et al (2011).
}
