% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sir.R
\name{compare_posteriors}
\alias{compare_posteriors}
\title{Function to compare posters}
\usage{
compare_posteriors(reference_sir = NULL, SIR, model_names = NULL,
  file_name = NULL)
}
\arguments{
\item{reference_sir}{Default = NULL, wether reference SIR fit}

\item{SIR}{list of SIR objects}

\item{model_names}{names of sir objects}

\item{file_name}{name of a file to identified the files exported by the
function. If NULL, does not save.}
}
\description{
Function to compare posters
}
