% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humpback_SIR_20171206.R
\name{HUMPBACK.SIR}
\alias{HUMPBACK.SIR}
\title{HUMPBACK SIR controls the sampling from the priors, the bisection and
likelihoods and the output functions}
\usage{
HUMPBACK.SIR(file_name = "NULL", n_resamples = 1000,
  priors = make_prior_list(),
  catch_multipliers = make_multiplier_list(),
  premodern_catch_multipliers = make_multiplier_list(),
  target.Yr = 2008, num.haplotypes = 66, output.Yrs = c(2008),
  abs.abundance = Abs.Abundance, rel.abundance = Rel.Abundance,
  rel.abundance.key = TRUE, count.data = NULL,
  count.data.key = FALSE, growth.rate.obs = c(0.074, 0.033, TRUE),
  growth.rate.Yrs = c(1995, 1996, 1997, 1998), catch.data = Catch.data,
  premodern_catch_data = NULL, realized_prior = FALSE,
  control = sir_control())
}
\arguments{
\item{file_name}{name of a file to identified the files exported by the
function}

\item{n_resamples}{number of resamples to compute the marginal posterior
distributions}

\item{priors}{List of priors, usually generated using \link{make_prior_list}.
Default is the default of \code{make_prior_list}. See details.}

\item{catch_multipliers}{List of catch multipliers, generated using \link{make_multiplier_list}
Can either be estimated or explicitly provided. Default is \code{make_multiplier_list}.}

\item{premodern_catch_multipliers}{List of historic catch multipliers, generated using \link{make_multiplier_list}
Can either be estimated or explicitly provided. Default is \code{make_multiplier_list}.}

\item{target.Yr}{year of the target population estimate for the bisection
method. Default is 2008}

\item{num.haplotypes}{number of haplotypes to compute minimum viable
population (from Jackson et al., 2006 and IWC, 2007)}

\item{output.Yrs}{year for outputing the predicted abundance estimates.
Default is 2008, but multiple years can be specified. For example, if
outputs for 2005 and 2008 are needed, output.Yrs = c(2005, 2008)}

\item{abs.abundance}{R object containing year, estimate of absolute
abundance, and CV (see example)}

\item{rel.abundance}{R object containing years, estimates of relative
abudnance and CVs (see example)}

\item{rel.abundance.key}{key to speficy if relative abundance data are used
in the likelihood. Default is TRUE}

\item{count.data}{R object containing years, estimates of counts and effort.
NOT USED}

\item{count.data.key}{key to speficy in count data are used. Default is
FALSE. NOT USED}

\item{growth.rate.obs}{observed growth rate (1st element) and standard error
(2nd element) as in Zerbni et al. (2011). If third element is FALSE, the
growth rate is not included in the likelihood}

\item{growth.rate.Yrs}{Years for which the growth.rate.obs were computed (as
in Zerbini et al., 2011)}

\item{catch.data}{R object containing the years and catches (see example)}

\item{premodern_catch_data}{R object containing the years maximum and minimum premodern catches}

\item{realized_prior}{Key to specify if realized prior is to be extracted. Default is FALSE.}

\item{control}{A list of control parameters, usually generated by
\code{sir_control}.}
}
\value{
A \code{list} containing posterior samples and metadata

TODO: Add the negative binomial likelihood for the count data, which is not
currently used even though it is defined in the main function call.

Current default prior specification:
\code{
make_prior_list(r_max = make_prior(runif, 0, 0.106),
                K = make_prior(use = FALSE),
                N_obs = make_prior(runif, 500, 20000),
                add_CV = make_prior(use = FALSE),
                z = make_prior(2.39),
                q_IA = make_prior(use = FALSE),
                q_count = make_prior(use = FALSE)

 make_multiplier_list(c_mult_1 = make_prior(1))}
}
\description{
HUMPBACK SIR controls the sampling from the priors, the bisection and
likelihoods and the output functions
}
