% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sir.R
\name{plot_density}
\alias{plot_density}
\title{OUTPUT FUNCTION}
\usage{
plot_density(SIR, file_name = "NULL", multiple_sirs = FALSE,
  lower = NULL, upper = NULL, prior_list = NULL,
  inc_reference = FALSE)
}
\arguments{
\item{SIR}{A fit SIR model or list of SIR models. Plots in the order provided.}

\item{file_name}{name of a file to identified the files exported by the
function.}

\item{multiple_sirs}{Logical whether or not multiple SIRS are provided as a list.}

\item{lower}{Vector of lower bounds for x-axis}

\item{upper}{Vector of upper bounds for x-axis}

\item{priors}{Default = NULL, wether realized priors are included in SIR}

\item{reference}{Default = NULL, wether reference case is included in SIR}
}
\value{
Returns and saves a figure with the posterior densities of parameters.
}
\description{
Function that provides a plot of the estimated posterior densities of parameters from  SIR  model.
}
